New World API
If we'd had a 1.0, this would be 2.0.

The old API was ambiguous and confusing. Use of `goTo` could easily lead
to unintentionally having the same path on the backstack twice. The
difference between `resetTo` and `replaceTo` was not easy to understand
and the two were easily confused.

Those three methods are all now deprecated (along with the previously
deprecated goUp), and there are two new methods to replace them:

- `set(Path)` will make the given path the top of your backstack,
  inferring push or pop operations and dispatching with the inferred
  Direction. This is your Just Go Here option.

- `set(Backstack, Direction)` will simply throw away the current
  backstack, install the given backstack, and dispatch with the given
  direction. This is your Total Control option.