Fix view state persistence
Thanks to Aleksey Malevaniy for catching these problems and
demonstrating fixes.

This change fixes two problems with view state persistence in
ActivityFlowSupport:

1. It was just throwing view state away when building the Parcelable to
save in the Bundle. The fix is to move filtering into Backstack itself,
so that it can build a Parcelable that includes view state.

2. It wasn (re)constructing the view hierarchy for the first time in
onResume(), which is after restoreHierarchyState() is called.  View
state was therefore not getting properly restored. The fix is to rebuild
the view hierarchy in onCreate(), so that it's there in time for
restoreHierarchyState().