Safe deprecation phase
flow.Path* is back, but entirely deprecated, with comments directing
users to flow.path.Path*.

Parceler is deprecated and replaced with StateParceler. Same purpose,
but works with Objects instead of Paths and has a more suitable name.

Backstack#current() is replaced with Backstack#top(). Again, Object
instead of Path + a better name.