Safer History building
State no longer lives in the History instance. This means building a
history is no longer unsafe. State is instead maintained by the
KeyManager, and History is just a list of keys.

States are created in the KeyManager on-demand to support traversals.

When the traversal queue is flushed we purge any stale/unused states
from the key manager.

Loading and saving state is no longer done via the History, but instead
handled directly by InternalLifecycleIntegration which pulls keys from
the current History and their States from the KeyManager; when loading,
it builds a History and pushes States into the KeyManager.

There are no longer any public APIs for loading and saving history, it's
all managed internally.

Fixes #148